Countries and CO's data
Random generation of preferred country and co for each player for the leaderboards table, the table now shows a country logo besides the players names.
